"Implement detailed help command and add help messages file"
This commit involves implementing a detailed help command in the application's discord command handler, to provide a more comprehensive and user-friendly guide for navigating the commands. A new 'helpMessages.json' file was also added to maintain the help content to support the help command. These changes aim to enhance the User Experience (UX) for better interaction with the app. Additionally, some unnecessary import was removed from 'discordUploader.ts' file and a second database data source was added for 'themes_database' in 'dataSources.xml'. This new data source adds support for working with a SQLite database.