Add AWS S3 SDK and JSON Web Token to package dependencies
Added the AWS SDK for JavaScript to the `package-lock.json` to allow interaction with AWS S3 service for file storage. In addition, JSON Web Token (jsonwebtoken) is added for implementing authentication in the application. The devDependencies section is also updated with related type definitions. These changes are made to support the upcoming features related to S3 object storage and user authentication.