Update joinable tester count and refactor project setup
Modified "getTotalTestersCount" in appstoreConnect.ts to "getJoinable" to fetch the accurate number of testers. The change was made to reflect the accurate count of testers who can join the application. Former function fetched total number of testers whereas new function fetches the remaining joinable spots.

Also, the discussions and unused code in "sendApproval.ts" are removed to improve code readability.

Added a new function in "startup.ts" to ensure required files and environment variables exist at start. This helps in setting up the project easier for new collaborators and reduces chances for runtime errors due to missing files or variables.

Formatting in bucketSetup.ts, cache.ts and main.ts was adjusted for better readability and maintenance. The S3 client in bucketSetup.ts is refactored to use environment variables directly.

Commented out code in testflight.ts is removed to declutter and improve readability.

Appstore connect headers are updated to fetch details in "getJoinable".

Removed unused imports and variables in different files to lighten the codebase and enhance maintainability.