Add theme fixing functionality in theme command
In this commit, functionality was added to the 'theme' command to fix any broken themes by comparing with a standard template and filling in any missing fields. This was facilitated through the creation of a new 'fix' subcommand. The 'jszip' module was also added to handle data for themes which are typically .zip files. The 'fixThemeCommand' function was created to implement this functionality.