Update attachment image file naming in discordUploader
Attachment image file naming for both light and dark themes in discordUploader.ts has been adjusted to include a "0-" prefix. This change is necessitated by a file naming convention alteration in the image storage system. Also, an unnecessary commented TODO section was removed in webserver.ts to declutter the code.