Update version number in package.json
This commit updates the version number of the application in the package.json file from version "1.0.2" to "1.1.2". This change is necessary to accurately represent the new functionalities that have been added in this release, such as the announcement feature. This version number increment will properly reflect the advancements made in the application since the last release.